+++ parsetree.go
@@ -22,6 +22,7 @@ const (
 	T_ICSTARTSWITH
 	T_ENDSWITH
 	T_ICENDSWITH
+	T_SHA1
 	T_NAME
 	T_PATH
 	T_STATS
@@ -139,6 +140,8 @@ func nodeString(treeNode *tnode) string {
 		return "ENDSWITH"
 	case T_ICENDSWITH:
 		return "ICENDSWITH"
+	case T_SHA1:
+		return "SHA1"
 	case T_NAME:
 		return "NAME"
 	case T_PATH: